# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESIDA API",
        "description": "REST API untuk sistem informasi manajemen warga perumahan. Mencakup manajemen penghuni, rumah, iuran, pembayaran, pengeluaran, dan laporan keuangan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Daftar pengeluaran (paginasi)",
                "description": "Daftar pengeluaran kas. Filter: kategori, tahun, bulan.",
                "operationId": "02dd0399645db746ef68a9bcdab69338",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter kategori",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "gaji_satpam",
                                "listrik_utilitas",
                                "kebersihan",
                                "perbaikan",
                                "lainnya"
                            ]
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter tahun",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Filter bulan (1-12)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Data per halaman (max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data pengeluaran."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Tambah pengeluaran baru",
                "description": "Mencatat pengeluaran kas baru.",
                "operationId": "c751fc8afec0e421ff4180beb6d02e87",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "description",
                                    "amount",
                                    "expense_date"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "listrik_utilitas",
                                        "enum": [
                                            "gaji_satpam",
                                            "listrik_utilitas",
                                            "kebersihan",
                                            "perbaikan",
                                            "lainnya"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bayar listrik Jan 2026",
                                        "maxLength": 255
                                    },
                                    "amount": {
                                        "description": "Harus > 0",
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengeluaran berhasil ditambahkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Detail pengeluaran",
                "description": "Detail pengeluaran berdasarkan ID.",
                "operationId": "816cba9dcf02e041a8b7a8070d498a6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengeluaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pengeluaran berhasil diambil."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Pengeluaran tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Update pengeluaran",
                "description": "Update data pengeluaran.",
                "operationId": "eb395073943129ebe9c47ddb07209b54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengeluaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "description",
                                    "amount",
                                    "expense_date"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "listrik_utilitas",
                                        "enum": [
                                            "gaji_satpam",
                                            "listrik_utilitas",
                                            "kebersihan",
                                            "perbaikan",
                                            "lainnya"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bayar listrik Jan 2026",
                                        "maxLength": 255
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data pengeluaran berhasil diperbarui."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Pengeluaran tidak ditemukan."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/expenses/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Nonaktifkan pengeluaran (soft delete)",
                "description": "Nonaktifkan pengeluaran (soft delete). Data tetap tersimpan.",
                "operationId": "d22cc8cb6388b7993f0e93267d8bb46c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengeluaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengeluaran berhasil dinonaktifkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Pengeluaran tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Sanctum. Format: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}